// src/libs/mrscraper-browser/firefox/server/types.ts
export type ProxySettings = {
  server: string;
  bypass?: string;
  username?: string;
  password?: string;
};